Conflicts:
	Ashikhmin/src/ModuleTests/ModuleTests.csproj
	Ashikhmin/src/ModuleTests/Properties/AssemblyInfo.cs
	Ashikhmin/src/OpaqueFunctions/FunctionAttributes.cs
	Ashikhmin/src/OpaqueFunctions/OpaqueFunctions.csproj
	Ashikhmin/src/OpaqueFunctions/Properties/AssemblyInfo.cs
	Ashikhmin/src/SummerPractice_2016.sln
	Ashikhmin/src/SummerPractice_2016/App.config
	Ashikhmin/src/SummerPractice_2016/Properties/AssemblyInfo.cs
	Ashikhmin/src/SummerPractice_2016/SummerPractice_2016.csproj
	Bakumenko/src/ModuleTests/ModuleTests.csproj
	Bakumenko/src/ModuleTests/Properties/AssemblyInfo.cs
	Bakumenko/src/OpaqueFunctions/FunctionAttributes.cs
	Bakumenko/src/OpaqueFunctions/OpaqueFunctions.csproj
	Bakumenko/src/OpaqueFunctions/Properties/AssemblyInfo.cs
	Bakumenko/src/SummerPractice_2016.sln
	Bakumenko/src/SummerPractice_2016/App.config
	Bakumenko/src/SummerPractice_2016/Properties/AssemblyInfo.cs
	Bakumenko/src/SummerPractice_2016/SummerPractice_2016.csproj
	Muradyan/TestingSystem/TestingSystem/gmffErrors.txt

+++ Lebedev/src/OpaqueFunctions/OpaqueFunctions.cs
@@ -1042,7 +1042,7 @@ namespace OpaqueFunctions
                 double F = 1;
                 for (int i = 1; i <= N; i++)
                 {
-                    F *= ((4 * i - 3) * Math.Pow(4 * i, -1)) * x;
+                    F *= ((4 * i - (double)3) * Math.Pow(4 * i, -1)) * x;
                     F = -F;
                     S += F;
                 }
